feat: processing of extracting dynamic nodes
